+++ CHANGELOG.md
@@ -41,6 +41,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## v4.5.6, unreleased
 
+### Added
+
+- Show more information during `wwctl container <shell|exec>` about when and if the container image will be rebuilt. #1302
+
 ### Fixed
 
 - Ensure autobuilt overlays include contextual overlay contents. #1296